♻️ Refactored chat sections.
